requirements.txt
streamlit
openai
openai
python-dotenv
torch
transformers
Pillow